Create ServiceBus common library.
