--- src/main/java/com/microsoft/azure/eventhubs/Exceptions/EntityNotFoundException.java
@@ -1,16 +1,0 @@
-package com.microsoft.azure.eventhubs.Exceptions;
-
-import com.microsoft.azure.eventhubs.common.RetryPolicy;
-
-public class EntityNotFoundException extends EventHubException 
-{
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = -1106827749824999989L;
-
-	@Override public boolean getIsTransient(){
-		  return false;
-	  }
-}
