--- src/main/java/com/microsoft/azure/eventhubs/RetryExponential.java
@@ -1,15 +1,0 @@
-package com.microsoft.azure.eventhubs;
-
-import java.util.Date;
-
-public final class RetryExponential extends RetryPolicy
-{
-	public RetryExponential(Date minimumBackoff, Date maximumBackoff, Date deltaBackoff){
-		
-	}
-	
-	// Will be removed: demo compilation purpose only
-	RetryExponential()
-	{
-	}
-}
