+++ src/main/java/com/microsoft/azure/servicebus/AuthorizationFailedException.java
@@ -1,0 +1,10 @@
+package com.microsoft.azure.servicebus;
+
+public class AuthorizationFailedException extends ServiceBusException {
+	private static final long serialVersionUID = -1106827749824999989L;
+	
+	@Override 
+	public boolean getIsTransient(){
+		  return false;
+	  }
+}
