+++ src/main/java/com/microsoft/azure/servicebus/ClientEntity.java
@@ -1,4 +1,4 @@
-package com.microsoft.azure.eventhubs.common;
+package com.microsoft.azure.servicebus;
 
 // contract for all client entities with Open-Close/Abort state m/c 
 public abstract class ClientEntity {
