+++ src/main/java/com/microsoft/azure/servicebus/EntityNotFoundException.java
@@ -1,0 +1,14 @@
+package com.microsoft.azure.servicebus;
+
+public class EntityNotFoundException extends ServiceBusException 
+{
+
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = -1106827749824999989L;
+
+	@Override public boolean getIsTransient(){
+		  return false;
+	  }
+}
