+++ src/main/java/com/microsoft/azure/servicebus/SendLinkHandler.java
@@ -1,4 +1,4 @@
-package com.microsoft.azure.eventhubs.common;
+package com.microsoft.azure.servicebus;
 
 import java.util.*;
 import java.util.concurrent.*;
