+++ src/main/java/com/microsoft/azure/servicebus/StringUtil.java
@@ -1,4 +1,4 @@
-package com.microsoft.azure.eventhubs.common;
+package com.microsoft.azure.servicebus;
 
 public final class StringUtil {
